+++ src/components/ExportCsvForm.tsx
@@ -1,9 +1,9 @@
 'use client';
-
-import { useActionState, useEffect, useRef } from 'react';
+import { useActionState } from 'react';
 import { useSelector } from 'react-redux';
 import { selectSelectedItems } from '@/store/selectedItemsSlice';
 import { exportCsvAction, ExportCsvState } from '@/app/actions/exportCsv';
+import { useCsvDownload } from '@/hooks/useCsvDownload';
 
 const initialState: ExportCsvState = { csvContent: null, error: null };
 
@@ -13,24 +13,8 @@ export const ExportCsvForm = () => {
     exportCsvAction,
     initialState
   );
-  const downloadedRef = useRef<string | null>(null);
 
-  useEffect(() => {
-    if (state.csvContent && state.csvContent !== downloadedRef.current) {
-      downloadedRef.current = state.csvContent;
-      const blob = new Blob([state.csvContent], {
-        type: 'text/csv;charset=utf-8',
-      });
-      const url = window.URL.createObjectURL(blob);
-      const a = document.createElement('a');
-      a.href = url;
-      a.download = 'selected-books.csv';
-      document.body.appendChild(a);
-      a.click();
-      window.URL.revokeObjectURL(url);
-      document.body.removeChild(a);
-    }
-  }, [state.csvContent]);
+  useCsvDownload(state.csvContent);
 
   if (selectedItems.length === 0) {
     return <p>No items selected for export</p>;
